Refactor: Split Design page into components
This commit refactors the Design page into smaller, reusable components to improve code organization and maintainability.
