+++ src/components/design/LoadingSpinner.tsx
@@ -1,0 +1,17 @@
+
+import React from 'react';
+
+interface LoadingSpinnerProps {
+  message?: string;
+}
+
+const LoadingSpinner = ({ message = "Loading..." }: LoadingSpinnerProps) => {
+  return (
+    <div className="flex justify-center items-center h-64">
+      <div className="animate-spin rounded-full h-12 w-12 border-t-2 border-b-2 border-blue-500"></div>
+      <span className="ml-3">{message}</span>
+    </div>
+  );
+};
+
+export default LoadingSpinner;
